Phase 1: pluggable storage backend (SQLite default, Postgres opt-in)
Splits upstream cache.py into a storage/ package selected at import time
by config.DATABASE_URL:

  * storage/sqlite_backend.py — near-verbatim move of upstream's SQLite
    logic. Default when DATABASE_URL is unset.
  * storage/postgres_backend.py — new. psycopg3 + ConnectionPool, mirrors
    the public API. Schema bootstrap serialised across replicas with a
    Postgres advisory lock so concurrent startup doesn't race.
  * storage/__init__.py — backend selector. Lazy-imports postgres_backend
    so SQLite-only deployments don't load psycopg.
  * cache.py — slim re-export shim. Every "from cache import …" callsite
    keeps working unchanged on either backend.

main.py gains a close_db() call in lifespan shutdown for graceful pool
drain.

Verified end-to-end:
  * SQLite mode: identical to upstream, all cache round-trips pass.
  * Postgres mode: all 6 tables created, upsert + dedup semantics match
    SQLite, /health returns 200, schema bootstrap reentrant.

Codex review: clean.
