+++ .env.example
@@ -23,6 +23,19 @@ MDBLIST_API_KEY=
 ACCESS_KEY=
 
 # =============================================================================
+# Storage backend (ElfHosted fork — hosted mode, opt-in)
+# =============================================================================
+
+# Leave DATABASE_URL unset to use SQLite (upstream default — recommended for
+# private/single-user deployments). Set it to a postgresql:// URL to switch
+# the cache layer to Postgres for multi-replica hosted deployments.
+#   Example:  DATABASE_URL=postgresql://postersplus:secret@postgres:5432/postersplus
+DATABASE_URL=
+# Connection pool sizing (only used when DATABASE_URL is set).
+DB_POOL_MIN_SIZE=1
+DB_POOL_MAX_SIZE=10
+
+# =============================================================================
 # Server
 # =============================================================================
