+++ requirements.txt
@@ -3,3 +3,8 @@ uvicorn
 Pillow
 httpx
 numpy
+# ElfHosted fork — Phase 1: Postgres backend (opt-in via DATABASE_URL).
+# Always installed so the image supports both private (SQLite) and hosted
+# (Postgres) modes without rebuilds. The postgres_backend module is only
+# imported when DATABASE_URL is set, so the dependency is otherwise inert.
+psycopg[binary,pool]>=3.1
